GC2D/CardLoad/titleDraw: 92.6% -> 94.0% (fix return signedness)

+++ src/GC2D/CardLoad.cpp
@@ -925,7 +925,7 @@ bool TCardLoad::titleDraw()
 		break;
 	}
 
-	return (u32)unk18 > 4;
+	return unk18 > 4;
 }
 
 void TCardLoad::makeBuffer(J2DTextBox* text_box, int size)
